requirements.txt
accelerate==1.10.1
addict==2.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
av==13.1.0
certifi==2025.8.3
charset-normalizer==3.4.0
colpali_engine==0.3.12
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
einops==0.8.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.19.1
fonttools==4.55.0
fsspec==2025.7.0
future==1.0.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
imageio==2.36.0
iopath==0.1.10
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==24.2
peft==0.16.0
pillow==11.3.0
platformdirs==4.3.6
portalocker==3.0.0
psutil==7.0.0
pydantic==2.11.5
pydantic_core==2.33.2
PyMuPDF==1.26.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.5
safetensors==0.4.5
SAM-2==1.0
scipy==1.15.3
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tifffile==2024.9.20
tokenizers==0.21.4
tomli==2.1.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.3
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yapf==0.43.0
